# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MuXiShooter",
        "title": "MuXiShooter",
        "termsOfService": "http://swagger.io/terms",
        "contact": {
            "name": "FireChickenMP4",
            "email": "13930176445@163.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/api/admin/get/getusers": {
            "get": {
                "description": "注: 管理员可用，查询结果是多重模糊搜索叠加的效果\n以及页码不输入或不合规范自动为第一页，每页多少不输入默认20，最多100\n如果查询结果不存在则返回切片为空\n用了id查询的话就一定只是一个确定的，而不是模糊搜索，其他参数就没用了（分页也是）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-user"
                ],
                "summary": "获取用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "权限组(user/admin)",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页多少，默认20，最大100",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PaginatedData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "登录状态异常",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "数据库查询失败",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
...
        "/api/admin/operation/deleteuser": {
            "delete": {
                "description": "管理员删除用户。ID=1(初始化管理员)不可删除；其他管理员仅可删除普通用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-user"
                ],
                "summary": "管理员删除用户",
                "parameters": [
                    {
                        "description": "删除用户请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdminDeleteUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "401": {
                        "description": "登录状态异常",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "数据库错误",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
...
        "/api/admin/update/usergroup": {
            "put": {
                "description": "ID=1(初始化管理员)权限组不可修改；仅ID=1可修改其他用户权限组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-user"
                ],
                "summary": "管理员修改用户权限组",
                "parameters": [
                    {
                        "description": "修改权限组请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdminUpdateUserGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "401": {
                        "description": "登录状态异常",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "数据库错误",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AdminDeleteUserRequest": {
            "description": "按用户ID删除用户",
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "dto.AdminUpdateUserGroupRequest": {
            "description": "按用户ID修改权限组，仅支持user/admin",
            "type": "object",
            "required": [
                "new_group",
                "user_id"
            ],
            "properties": {
                "new_group": {
                    "type": "string",
                    "enum": [
                        "user",
                        "admin"
                    ]
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.PaginatedData": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "查询所得列表"
                },
                "page": {
                    "description": "页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页多少",
                    "type": "integer"
                },
                "total": {
                    "description": "返回数据总数",
                    "type": "integer"
                }
            }
        },
...
        "dto.Response": {
            "description": "通用响应结构体",
            "type": "object",
            "properties": {
                "code": {
                    "description": "http状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "根据具体数据来定类型"
                },
                "message": {
                    "description": "返回的消息",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入你的Bearer Token，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}